refactor: hoist Azure Key Vault wiring into shared lib
Move Options/KeyVaultOptions.cs into Portals.Shared.Configuration and add KeyVaultConfigurationExtensions.AddPortalsKeyVault to encapsulate the DefaultAzureCredential wiring. Program.cs collapses 25+ lines of inline KeyVault setup into one call. The Azure.Extensions.AspNetCore.Configuration.Secrets and Azure.Identity NuGet refs move to Portals.Umbraco.Shared.csproj.
