+++ apps/umbraco-infoportal/Program.cs
@@ -1,7 +1,5 @@
-using Azure.Extensions.AspNetCore.Configuration.Secrets;
-using Azure.Identity;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
-using umbraco_infoportal.Options;
+using Portals.Shared.Configuration;
 using Umbraco.Cms.Core.Composing;
 
 
@@ -9,34 +7,10 @@ WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
 builder.WebHost.ConfigureKestrel(options =>
 {
-    options.Limits.MaxRequestHeadersTotalSize = 65536; 
+    options.Limits.MaxRequestHeadersTotalSize = 65536;
 });
 
-builder.Services.Configure<KeyVaultOptions>(
-    builder.Configuration.GetSection(KeyVaultOptions.SectionName));
-
-KeyVaultOptions keyVaultOptions = builder.Configuration
-    .GetSection(KeyVaultOptions.SectionName)
-    .Get<KeyVaultOptions>() ?? new();
-
-bool keyVaultEnabled = keyVaultOptions.Enabled ?? !builder.Environment.IsDevelopment();
-
-if (keyVaultEnabled)
-{
-    if (string.IsNullOrWhiteSpace(keyVaultOptions.AkvUri))
-    {
-        throw new InvalidOperationException(
-            $"Configuration value '{KeyVaultOptions.AkvUriKey}' must be configured when Key Vault is enabled.");
-    }
-
-    if (!Uri.TryCreate(keyVaultOptions.AkvUri, UriKind.Absolute, out Uri? keyVaultEndpoint))
-    {
-        throw new InvalidOperationException(
-            $"Configuration value '{KeyVaultOptions.AkvUriKey}' must be an absolute URI.");
-    }
-
-    builder.Configuration.AddAzureKeyVault(keyVaultEndpoint, new DefaultAzureCredential());
-}
+builder.Configuration.AddPortalsKeyVault(builder.Configuration, builder.Environment);
 
 builder.CreateUmbracoBuilder()
     .AddBackOffice()
