+++ apps/umbraco-infoportal/umbraco-infoportal.csproj
@@ -8,8 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Azure.Extensions.AspNetCore.Configuration.Secrets" Version="1.5.0" />
-    <PackageReference Include="Azure.Identity" Version="1.21.0" />
     <PackageReference Include="MimeKit" Version="4.15.1" />
     <PackageReference Include="SQLitePCLRaw.bundle_e_sqlite3" Version="3.0.2" />
     <PackageReference Include="Umbraco.Cms" Version="17.3.4" />
